:sparkles: Created a separate file for API and created combine reducers and made changes based on that in the files
