+++ package.json
@@ -17,6 +17,7 @@
     "react-redux": "^8.0.5",
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
+    "redux-toolkit": "^1.1.2",
     "styled-components": "^6.0.0-rc.1"
   },
   "devDependencies": {
