--- src/action/userAction.jsx
@@ -1,37 +1,0 @@
-import axios from "axios";
-
-// Action Types
-export const FETCH_USERS_REQUEST = "FETCH_USERS_REQUEST";
-export const FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS";
-export const FETCH_USERS_FAILURE = "FETCH_USERS_FAILURE";
-
-// Action Creators
-export const fetchUsersRequest = () => ({
-  type: FETCH_USERS_REQUEST,
-});
-
-export const fetchUsersSuccess = (users) => ({
-  type: FETCH_USERS_SUCCESS,
-  payload: users,
-});
-
-export const fetchUsersFailure = (error) => ({
-  type: FETCH_USERS_FAILURE,
-  payload: error,
-});
-
-// Async Action
-export const fetchUsers = (page) => {
-  return (dispatch) => {
-    dispatch(fetchUsersRequest());
-    axios
-      .get(`https://reqres.in/api/users?page=${page}`)
-      .then((response) => {
-        const users = response.data.data;
-        dispatch(fetchUsersSuccess(users));
-      })
-      .catch((error) => {
-        dispatch(fetchUsersFailure(error.message));
-      });
-  };
-};
