+++ src/main.jsx
@@ -5,9 +5,9 @@ import GlobalStyle from "@src/styled/GlobalStyle.style";
 import { Provider } from "react-redux";
 import { createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
-import userReducer from "./reducer/userReducer.jsx";
+import { rootReducer } from "./reducer/userReducer.js";
 
-const store = createStore(userReducer, applyMiddleware(thunk));
+const store = createStore(rootReducer, applyMiddleware(thunk));
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
