--- src/reducer/userReducer.jsx
@@ -1,40 +1,0 @@
-import {
-  FETCH_USERS_REQUEST,
-  FETCH_USERS_SUCCESS,
-  FETCH_USERS_FAILURE,
-} from "../action/userAction";
-
-// Initial state
-const initialState = {
-  loading: false,
-  users: [],
-  error: "",
-};
-
-// Reducer function
-const userReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case FETCH_USERS_REQUEST:
-      return {
-        ...state,
-        loading: true,
-        error: "",
-      };
-    case FETCH_USERS_SUCCESS:
-      return {
-        ...state,
-        loading: false,
-        users: action.payload,
-      };
-    case FETCH_USERS_FAILURE:
-      return {
-        ...state,
-        loading: false,
-        error: action.payload,
-      };
-    default:
-      return state;
-  }
-};
-
-export default userReducer;
