--- src/store/store.jsx
@@ -1,6 +1,0 @@
-import { createStore } from "redux";
-import userReducer from "@src/reducer/userReducer";
-
-const store = createStore(userReducer);
-
-export default store;
